+++ code/preprocessing/preprocessing.hpp
@@ -19,11 +19,32 @@ void eliminated_facts_removal(hplus::instance& inst, hplus::statistics& stats, s
 
 void eliminated_actions_removal(hplus::instance& inst, hplus::statistics& stats);
 
+void lmcut_landmarks_extraction(hplus::instance& inst);
+
 inline void prepare_preprocessing(hplus::instance& inst) {
     inst.fixed_facts = binary_set{inst.n};
     inst.fixed_actions = binary_set{inst.m};
     inst.eliminated_facts = binary_set{inst.n};
     inst.eliminated_actions = binary_set{inst.m};
+}
+
+inline void prepare_optimization_helpers(hplus::instance& inst) {
+    inst.fadd_cpx_start.clear();
+    inst.fadd_cpx_start.reserve(inst.m);
+    unsigned int sum{0};
+    for (const auto& act : inst.actions) {
+        inst.fadd_cpx_start.push_back(sum);
+        sum += act.eff_sparse.size();
+    }
+    inst.act_with_pre = std::vector<std::vector<unsigned int>>(inst.n);
+    inst.act_with_eff = std::vector<std::vector<unsigned int>>(inst.n);
+    for (unsigned int act_i = 0; act_i < inst.m; ++act_i) {
+        for (const auto& pre_i : inst.actions[act_i].pre_sparse) inst.act_with_pre[pre_i].push_back(act_i);
+        for (const auto& eff_i : inst.actions[act_i].eff_sparse) inst.act_with_eff[eff_i].push_back(act_i);
+    }
+
+    inst.eliminated_facts = binary_set{1};
+    inst.eliminated_actions = binary_set{1};
 }
 
 inline void preprocess(const hplus::execution& exec, hplus::instance& inst, hplus::statistics& stats) {
@@ -43,34 +64,11 @@ inline void preprocess(const hplus::execution& exec, hplus::instance& inst, hplu
     dominated_actions_extraction(inst, landmarks);
     eliminated_actions_removal(inst, stats);
 
+    prepare_optimization_helpers(inst);
+
+    lmcut_landmarks_extraction(inst);
+
     stats.preprocessing = GET_TIME() - start_time;
-}
-
-inline void prepare_optimization_helpers(const hplus::execution& exec, hplus::instance& inst, hplus::statistics& stats) {
-    if (BASIC_VERBOSE()) LOG_INFO << "Preparing optimization helpers";
-
-    double start_time = GET_TIME();
-    double prep_time_save = stats.preprocessing;
-    stats.preprocessing += static_cast<double>(exec.timelimit) - start_time;
-
-    inst.fadd_cpx_start.clear();
-    inst.fadd_cpx_start.reserve(inst.m);
-    unsigned int sum{0};
-    for (const auto& act : inst.actions) {
-        inst.fadd_cpx_start.push_back(sum);
-        sum += act.eff_sparse.size();
-    }
-    inst.act_with_pre = std::vector<std::vector<unsigned int>>(inst.n);
-    inst.act_with_eff = std::vector<std::vector<unsigned int>>(inst.n);
-    for (unsigned int act_i = 0; act_i < inst.m; ++act_i) {
-        for (const auto& pre_i : inst.actions[act_i].pre_sparse) inst.act_with_pre[pre_i].push_back(act_i);
-        for (const auto& eff_i : inst.actions[act_i].eff_sparse) inst.act_with_eff[eff_i].push_back(act_i);
-    }
-
-    inst.eliminated_facts = binary_set{1};
-    inst.eliminated_actions = binary_set{1};
-
-    stats.preprocessing = prep_time_save + GET_TIME() - start_time;
 }
 
 }  // namespace prep